[refactor] #47 OpenAi 호출 관련 코드들을 PossibilityCommandService에서 OpenAiRecommendationClient로 이동
